Refactor: move scan components to their own files
This creates separate files for the components:

* aggregator
* pki
* resultserver
* scan

Making the code slightly easier to navigate, than having to read one big
file.
