+++ pkg/controller/compliancescan/compliancescan_controller.go
@@ -2,12 +2,9 @@ package compliancescan
 
 import (
 	"context"
-	"fmt"
-	"path"
 	"time"
 
 	"github.com/go-logr/logr"
-	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -25,7 +22,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
 	complianceoperatorv1alpha1 "github.com/openshift/compliance-operator/pkg/apis/complianceoperator/v1alpha1"
-	"github.com/openshift/compliance-operator/pkg/controller/common"
 )
 
 var log = logf.Log.WithName("controller_compliancescan")
@@ -375,119 +371,6 @@ func (r *ReconcileComplianceScan) phaseDoneHandler(instance *complianceoperatorv
 	return reconcile.Result{}, nil
 }
 
-func (r *ReconcileComplianceScan) handleRootCASecret(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	exist, err := secretExists(r.client, RootCAPrefix+instance.Name, common.GetComplianceOperatorNamespace())
-	if err != nil {
-		return err
-	}
-	if exist {
-		return nil
-	}
-
-	logger.Info("creating CA", "instance", instance.Name)
-	secret, err := makeCASecret(instance, common.GetComplianceOperatorNamespace())
-	if err != nil {
-		return err
-	}
-
-	if err = controllerutil.SetControllerReference(instance, secret, r.scheme); err != nil {
-		return err
-	}
-
-	// Create the CA secret.
-	err = r.client.Create(context.TODO(), secret)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	return nil
-}
-
-func (r *ReconcileComplianceScan) handleResultServerSecret(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	exist, err := secretExists(r.client, ServerCertPrefix+instance.Name, common.GetComplianceOperatorNamespace())
-	if err != nil {
-		return err
-	}
-	if exist {
-		return nil
-	}
-
-	logger.Info("creating server cert", "instance", instance.Name)
-	secret, err := makeServerCertSecret(r.client, instance, common.GetComplianceOperatorNamespace())
-	if err != nil {
-		return err
-	}
-
-	if err = controllerutil.SetControllerReference(instance, secret, r.scheme); err != nil {
-		return err
-	}
-
-	// Create the server cert secret.
-	err = r.client.Create(context.TODO(), secret)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-	return nil
-}
-
-func (r *ReconcileComplianceScan) handleResultClientSecret(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	exist, err := secretExists(r.client, ClientCertPrefix+instance.Name, common.GetComplianceOperatorNamespace())
-	if err != nil {
-		return err
-	}
-	if exist {
-		return nil
-	}
-
-	logger.Info("creating client cert", "instance", instance.Name)
-	secret, err := makeClientCertSecret(r.client, instance, common.GetComplianceOperatorNamespace())
-	if err != nil {
-		return err
-	}
-
-	if err = controllerutil.SetControllerReference(instance, secret, r.scheme); err != nil {
-		return err
-	}
-
-	// Create the client cert secret.
-	err = r.client.Create(context.TODO(), secret)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	return nil
-}
-
-func (r *ReconcileComplianceScan) deleteRootCASecret(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	logger.Info("deleting CA", "instance", instance.Name)
-	ns := common.GetComplianceOperatorNamespace()
-	secret := certSecret(getCASecretName(instance), ns, []byte{}, []byte{}, []byte{})
-	return r.deleteSecret(secret)
-}
-
-func (r *ReconcileComplianceScan) deleteResultServerSecret(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	logger.Info("deleting server cert", "instance", instance.Name)
-	ns := common.GetComplianceOperatorNamespace()
-	secret := certSecret(getServerCertSecretName(instance), ns, []byte{}, []byte{}, []byte{})
-	return r.deleteSecret(secret)
-}
-
-func (r *ReconcileComplianceScan) deleteResultClientSecret(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	logger.Info("deleting client cert", "instance", instance.Name)
-	ns := common.GetComplianceOperatorNamespace()
-	secret := certSecret(getClientCertSecretName(instance), ns, []byte{}, []byte{}, []byte{})
-	return r.deleteSecret(secret)
-}
-
-func (r *ReconcileComplianceScan) deleteSecret(secret *corev1.Secret) error {
-	// Delete the client cert secret.
-	err := r.client.Delete(context.TODO(), secret)
-	if err != nil && !errors.IsNotFound(err) {
-		return err
-	}
-	return nil
-}
-
 func getTargetNodes(r *ReconcileComplianceScan, instance *complianceoperatorv1alpha1.ComplianceScan) (corev1.NodeList, error) {
 	var nodes corev1.NodeList
 
@@ -500,132 +383,6 @@ func getTargetNodes(r *ReconcileComplianceScan, instance *complianceoperatorv1al
 	}
 
 	return nodes, nil
-}
-
-func (r *ReconcileComplianceScan) createScanPods(instance *complianceoperatorv1alpha1.ComplianceScan, nodes corev1.NodeList, logger logr.Logger) error {
-	// On each eligible node..
-	for _, node := range nodes.Items {
-		// ..schedule a pod..
-		logger.Info("Creating a pod for node", "node", node.Name)
-		pod := newScanPodForNode(instance, &node, logger)
-		if err := controllerutil.SetControllerReference(instance, pod, r.scheme); err != nil {
-			log.Error(err, "Failed to set pod ownership", "pod", pod)
-			return err
-		}
-
-		// ..and launch it..
-		err := r.client.Create(context.TODO(), pod)
-		if errors.IsAlreadyExists(err) {
-			logger.Info("Pod already exists. This is fine.", "pod", pod)
-		} else if err != nil {
-			log.Error(err, "Failed to launch a pod", "pod", pod)
-			return err
-		} else {
-			logger.Info("Launched a pod", "pod", pod)
-			// ..since the pod name can be random, store it in a label
-			setPodForNodeName(instance, node.Name, pod.Name)
-		}
-	}
-
-	// make sure the instance is updated with the node-pod labels
-	if err := r.client.Update(context.TODO(), instance); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (r *ReconcileComplianceScan) deleteScanPods(instance *complianceoperatorv1alpha1.ComplianceScan, nodes corev1.NodeList, logger logr.Logger) error {
-	// On each eligible node..
-	for _, node := range nodes.Items {
-		// ..schedule a pod..
-		logger.Info("Creating a pod for node", "node", node.Name)
-		pod := newScanPodForNode(instance, &node, logger)
-
-		// ..and launch it..
-		err := r.client.Delete(context.TODO(), pod)
-		if errors.IsNotFound(err) {
-			logger.Info("Pod is already gone. This is fine.", "pod", pod)
-		} else if err != nil {
-			log.Error(err, "Failed to delete a pod", "pod", pod)
-			return err
-		} else {
-			logger.Info("deleted pod", "pod", pod)
-		}
-	}
-
-	return nil
-}
-
-// The result-server is a pod that listens for results from other pods and
-// stores them in a PVC.
-// It's comprised of the PVC for the scan, the pod and a service that fronts it
-func (r *ReconcileComplianceScan) createResultServer(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	err := r.createPVCForScan(instance)
-	if err != nil {
-		logger.Error(err, "Cannot create the PersistentVolumeClaims")
-		return err
-	}
-
-	resultServerLabels := map[string]string{
-		"complianceScan": instance.Name,
-		"app":            "resultserver",
-	}
-
-	logger.Info("Creating scan result server pod")
-	deployment := resultServer(instance, resultServerLabels, logger)
-	if err = controllerutil.SetControllerReference(instance, deployment, r.scheme); err != nil {
-		log.Error(err, "Failed to set deployment ownership", "deployment", deployment)
-		return err
-	}
-
-	err = r.client.Create(context.TODO(), deployment)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		logger.Error(err, "Cannot create deployment", "deployment", deployment)
-		return err
-	}
-	logger.Info("ResultServer Deployment launched", "name", deployment.Name)
-
-	service := resultServerService(instance, resultServerLabels)
-	if err = controllerutil.SetControllerReference(instance, service, r.scheme); err != nil {
-		log.Error(err, "Failed to set service ownership", "service", service)
-		return err
-	}
-
-	err = r.client.Create(context.TODO(), service)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		logger.Error(err, "Cannot create service", "service", service)
-		return err
-	}
-	logger.Info("ResultServer Service launched", "name", service.Name)
-	return nil
-}
-
-func (r *ReconcileComplianceScan) deleteResultServer(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	resultServerLabels := map[string]string{
-		"complianceScan": instance.Name,
-		"app":            "resultserver",
-	}
-
-	logger.Info("Deleting scan result server pod")
-
-	deployment := resultServer(instance, resultServerLabels, logger)
-
-	err := r.client.Delete(context.TODO(), deployment)
-	if err != nil && !errors.IsNotFound(err) {
-		logger.Error(err, "Cannot delete deployment", "deployment", deployment)
-		return err
-	}
-	logger.Info("ResultServer Deployment deleted", "name", deployment.Name)
-	logger.Info("Deleting scan result server service")
-
-	service := resultServerService(instance, resultServerLabels)
-	err = r.client.Delete(context.TODO(), service)
-	if err != nil && !errors.IsNotFound(err) {
-		logger.Error(err, "Cannot delete service", "service", service)
-		return err
-	}
-	logger.Info("ResultServer Service deleted", "name", service.Name)
-	return nil
 }
 
 func (r *ReconcileComplianceScan) createPVCForScan(instance *complianceoperatorv1alpha1.ComplianceScan) error {
@@ -662,25 +419,6 @@ func isPodRunning(r *ReconcileComplianceScan, podName, namespace string, logger 
 	// the pod is still running or being created etc
 	logger.Info("Pod still running", "pod", podName)
 	return true, nil
-}
-
-func getScanResult(cm *corev1.ConfigMap) (complianceoperatorv1alpha1.ComplianceScanStatusResult, error) {
-	exitcode, ok := cm.Data["exit-code"]
-	if ok {
-		switch exitcode {
-		case "0":
-			return complianceoperatorv1alpha1.ResultCompliant, nil
-		case "2":
-			return complianceoperatorv1alpha1.ResultNonCompliant, nil
-		default:
-			errorMsg, ok := cm.Data["error-msg"]
-			if ok {
-				return complianceoperatorv1alpha1.ResultError, fmt.Errorf(errorMsg)
-			}
-			return complianceoperatorv1alpha1.ResultError, fmt.Errorf("The ConfigMap '%s' was missing 'error-msg'", cm.Name)
-		}
-	}
-	return complianceoperatorv1alpha1.ResultError, fmt.Errorf("The ConfigMap '%s' was missing 'exit-code'", cm.Name)
 }
 
 // gatherResults will iterate the nodes in the scan and get the results
@@ -756,257 +494,6 @@ func getPVCForScan(instance *complianceoperatorv1alpha1.ComplianceScan) *corev1.
 	}
 }
 
-func newScanPodForNode(scanInstance *complianceoperatorv1alpha1.ComplianceScan, node *corev1.Node, logger logr.Logger) *corev1.Pod {
-
-	mode := int32(0744)
-
-	podName := createPodForNodeName(scanInstance.Name, node.Name)
-	cmName := getConfigMapForNodeName(scanInstance.Name, node.Name)
-	podLabels := map[string]string{
-		"complianceScan": scanInstance.Name,
-		"targetNode":     node.Name,
-	}
-
-	return &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      podName,
-			Namespace: scanInstance.Namespace,
-			Labels:    podLabels,
-		},
-		Spec: corev1.PodSpec{
-			ServiceAccountName: "compliance-operator",
-			InitContainers: []corev1.Container{
-				{
-					Name:  "content-container",
-					Image: getInitContainerImage(&scanInstance.Spec, logger),
-					Command: []string{
-						"sh",
-						"-c",
-						fmt.Sprintf("cp %s /content | /bin/true", path.Join("/", scanInstance.Spec.Content)),
-					},
-					ImagePullPolicy: corev1.PullAlways,
-					VolumeMounts: []corev1.VolumeMount{
-						{
-							Name:      "content-dir",
-							MountPath: "/content",
-						},
-					},
-				},
-			},
-			Containers: []corev1.Container{
-				{
-					Name:  "log-collector",
-					Image: GetComponentImage(LOG_COLLECTOR),
-					Args: []string{
-						"--arf-file=/reports/report-arf.xml",
-						"--results-file=/reports/report.xml",
-						"--exit-code-file=/reports/exit_code",
-						"--oscap-output-file=/reports/cmd_output",
-						"--config-map-name=" + cmName,
-						"--owner=" + scanInstance.Name,
-						"--namespace=" + scanInstance.Namespace,
-						"--resultserveruri=" + getResultServerURI(scanInstance),
-						"--tls-client-cert=/etc/pki/tls/tls.crt",
-						"--tls-client-key=/etc/pki/tls/tls.key",
-						"--tls-ca=/etc/pki/tls/ca.crt",
-					},
-					SecurityContext: &corev1.SecurityContext{
-						Privileged: &trueVal,
-					},
-					ImagePullPolicy: corev1.PullAlways,
-					VolumeMounts: []corev1.VolumeMount{
-						{
-							Name:      "report-dir",
-							MountPath: "/reports",
-							ReadOnly:  true,
-						},
-						{
-							Name:      "tls",
-							MountPath: "/etc/pki/tls",
-						},
-					},
-				},
-				{
-					Name:    OpenSCAPScanContainerName,
-					Image:   GetComponentImage(OPENSCAP),
-					Command: []string{OpenScapScriptPath},
-					SecurityContext: &corev1.SecurityContext{
-						Privileged: &trueVal,
-					},
-					VolumeMounts: []corev1.VolumeMount{
-						{
-							Name:      "host",
-							MountPath: "/host",
-						},
-						{
-							Name:      "report-dir",
-							MountPath: "/reports",
-						},
-						{
-							Name:      "content-dir",
-							MountPath: "/content",
-						},
-						{
-							Name:      scanInstance.Labels[OpenSCAPScriptCmLabel],
-							MountPath: "/scripts",
-						},
-					},
-					EnvFrom: []corev1.EnvFromSource{
-						{
-							ConfigMapRef: &corev1.ConfigMapEnvSource{
-								LocalObjectReference: corev1.LocalObjectReference{
-									Name: scanInstance.Labels[OpenSCAPScriptEnvLabel],
-								},
-							},
-						},
-					},
-				},
-			},
-			Tolerations: []corev1.Toleration{
-				{
-					Key:      "node-role.kubernetes.io/master",
-					Operator: "Exists",
-					Effect:   "NoSchedule",
-				},
-			},
-			NodeSelector: map[string]string{
-				NodeHostnameLabel: node.Labels[NodeHostnameLabel],
-			},
-			RestartPolicy: corev1.RestartPolicyOnFailure,
-			Volumes: []corev1.Volume{
-				{
-					Name: "host",
-					VolumeSource: corev1.VolumeSource{
-						HostPath: &corev1.HostPathVolumeSource{
-							Path: "/",
-							Type: &hostPathDir,
-						},
-					},
-				},
-				{
-					Name: "report-dir",
-					VolumeSource: corev1.VolumeSource{
-						EmptyDir: &corev1.EmptyDirVolumeSource{},
-					},
-				},
-				{
-					Name: "content-dir",
-					VolumeSource: corev1.VolumeSource{
-						EmptyDir: &corev1.EmptyDirVolumeSource{},
-					},
-				},
-				{
-					Name: scanInstance.Labels[OpenSCAPScriptCmLabel],
-					VolumeSource: corev1.VolumeSource{
-						ConfigMap: &corev1.ConfigMapVolumeSource{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: scanInstance.Labels[OpenSCAPScriptCmLabel],
-							},
-							DefaultMode: &mode,
-						},
-					},
-				},
-				{
-					Name: "tls",
-					VolumeSource: corev1.VolumeSource{
-						Secret: &corev1.SecretVolumeSource{
-							SecretName: ClientCertPrefix + scanInstance.Name,
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-// Serve up arf reports for a compliance scan with a web service protected by openshift auth (oauth-proxy sidecar).
-// Needs corresponding Service (with service-serving cert).
-// Need to aggregate reports into one service ? on subdirs?
-func resultServer(scanInstance *complianceoperatorv1alpha1.ComplianceScan, labels map[string]string, logger logr.Logger) *appsv1.Deployment {
-	return &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      getResultServerName(scanInstance),
-			Namespace: scanInstance.Namespace,
-		},
-		Spec: appsv1.DeploymentSpec{
-			Replicas: &oneReplica,
-			Selector: &metav1.LabelSelector{
-				MatchLabels: labels,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
-				},
-				Spec: corev1.PodSpec{
-					// TODO(jaosorior): Should we schedule this in the master nodes only?
-					ServiceAccountName: "compliance-operator",
-					Containers: []corev1.Container{
-						{
-							Name:            "result-server",
-							Image:           GetComponentImage(RESULT_SERVER),
-							ImagePullPolicy: corev1.PullAlways,
-							Args: []string{
-								"--path=/reports/",
-								"--address=0.0.0.0",
-								fmt.Sprintf("--port=%d", ResultServerPort),
-								"--tls-server-cert=/etc/pki/tls/tls.crt",
-								"--tls-server-key=/etc/pki/tls/tls.key",
-								"--tls-ca=/etc/pki/tls/ca.crt",
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									Name:      "arfreports",
-									MountPath: "/reports",
-								},
-								{
-									Name:      "tls",
-									MountPath: "/etc/pki/tls",
-								},
-							},
-						},
-					},
-					Volumes: []corev1.Volume{
-						{
-							Name: "arfreports",
-							VolumeSource: corev1.VolumeSource{
-								PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-									ClaimName: getPVCForScanName(scanInstance),
-								},
-							},
-						},
-						{
-							Name: "tls",
-							VolumeSource: corev1.VolumeSource{
-								Secret: &corev1.SecretVolumeSource{
-									SecretName: ServerCertPrefix + scanInstance.Name,
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func resultServerService(scanInstance *complianceoperatorv1alpha1.ComplianceScan, labels map[string]string) *corev1.Service {
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      getResultServerName(scanInstance),
-			Namespace: scanInstance.Namespace,
-		},
-		Spec: corev1.ServiceSpec{
-			Selector: labels,
-			Ports: []corev1.ServicePort{
-				{
-					Protocol: corev1.Protocol("TCP"),
-					Port:     ResultServerPort,
-				},
-			},
-		},
-	}
-}
-
 // pod names are limited to 63 chars, inclusive. Try to use a friendly name, if that can't be done,
 // just use a hash. Either way, the node would be present in a label of the pod.
 func createPodForNodeName(scanName, nodeName string) string {
@@ -1039,14 +526,6 @@ func getPVCForScanName(instance *complianceoperatorv1alpha1.ComplianceScan) stri
 	return instance.Name
 }
 
-func getResultServerName(instance *complianceoperatorv1alpha1.ComplianceScan) string {
-	return instance.Name + "-rs"
-}
-
-func getResultServerURI(instance *complianceoperatorv1alpha1.ComplianceScan) string {
-	return "https://" + getResultServerName(instance) + fmt.Sprintf(":%d/", ResultServerPort)
-}
-
 func getInitContainerImage(scanSpec *complianceoperatorv1alpha1.ComplianceScanSpec, logger logr.Logger) string {
 	image := DefaultContentContainerImage
 
@@ -1056,112 +535,4 @@ func getInitContainerImage(scanSpec *complianceoperatorv1alpha1.ComplianceScanSp
 
 	logger.Info("Content image", "image", image)
 	return image
-}
-
-func createAggregatorPodName(scanName string) string {
-	return dnsLengthName("aggregator-pod-", "aggregator-pod-%s", scanName)
-}
-
-func newAggregatorPod(scanInstance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) *corev1.Pod {
-	podName := createAggregatorPodName(scanInstance.Name)
-
-	podLabels := map[string]string{
-		"complianceScan": scanInstance.Name,
-	}
-
-	return &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      podName,
-			Namespace: scanInstance.Namespace,
-			Labels:    podLabels,
-		},
-		Spec: corev1.PodSpec{
-			ServiceAccountName: "compliance-operator",
-			InitContainers: []corev1.Container{
-				{
-					Name:  "content-container",
-					Image: getInitContainerImage(&scanInstance.Spec, logger),
-					Command: []string{
-						"sh",
-						"-c",
-						fmt.Sprintf("cp %s /content | /bin/true", path.Join("/", scanInstance.Spec.Content)),
-					},
-					ImagePullPolicy: corev1.PullAlways,
-					VolumeMounts: []corev1.VolumeMount{
-						{
-							Name:      "content-dir",
-							MountPath: "/content",
-						},
-					},
-				},
-			},
-			Containers: []corev1.Container{
-				{
-					Name:  "log-collector",
-					Image: GetComponentImage(AGGREGATOR),
-					Args: []string{
-						"--content=" + absContentPath(scanInstance.Spec.Content),
-						"--scan=" + scanInstance.Name,
-					},
-					SecurityContext: &corev1.SecurityContext{
-						Privileged: &trueVal,
-					},
-					VolumeMounts: []corev1.VolumeMount{
-						{
-							Name:      "content-dir",
-							MountPath: "/content",
-						},
-					},
-				},
-			},
-			RestartPolicy: corev1.RestartPolicyOnFailure,
-			Volumes: []corev1.Volume{
-				{
-					Name: "content-dir",
-					VolumeSource: corev1.VolumeSource{
-						EmptyDir: &corev1.EmptyDirVolumeSource{},
-					},
-				},
-			},
-		},
-	}
-}
-
-func (r *ReconcileComplianceScan) launchAggregatorPod(scanInstance *complianceoperatorv1alpha1.ComplianceScan, pod *corev1.Pod, logger logr.Logger) error {
-	// Make use of optimistic concurrency and just try creating the pod
-	err := r.client.Create(context.TODO(), pod)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		logger.Error(err, "Cannot launch pod", "pod", pod)
-		return err
-	}
-
-	if errors.IsAlreadyExists(err) {
-		// If the pod was already created, just return
-		return nil
-	}
-
-	if scanInstance.Annotations == nil {
-		scanInstance.Annotations = make(map[string]string)
-	}
-
-	scanInstance.Annotations[AggregatorPodAnnotation] = pod.Name
-	return r.client.Update(context.TODO(), scanInstance)
-}
-
-func (r *ReconcileComplianceScan) deleteAggregator(instance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) error {
-	aggregator := newAggregatorPod(instance, logger)
-	err := r.client.Delete(context.TODO(), aggregator)
-	if err != nil && !errors.IsNotFound(err) {
-		logger.Error(err, "Cannot delete aggregator pod", "pod", aggregator)
-		return err
-	}
-
-	return nil
-}
-
-func isAggregatorRunning(r *ReconcileComplianceScan, scanInstance *complianceoperatorv1alpha1.ComplianceScan, logger logr.Logger) (bool, error) {
-	logger.Info("Checking aggregator pod for scan", "scan", scanInstance.Name)
-
-	podName := scanInstance.Annotations[AggregatorPodAnnotation]
-	return isPodRunning(r, podName, scanInstance.Namespace, logger)
 }
